faire les roles et les permisson pour lentite produit

+++ app/Http/Controllers/TenueController.php
@@ -2,7 +2,7 @@
 namespace App\Http\Controllers;
 
 use App\Models\Tenue;
-use App\Models\Categorie;
+use App\Models\Category;
 use App\Http\Requests\TenueRequest;
 
 use Illuminate\Http\Request;
@@ -11,18 +11,28 @@ class TenueController extends Controller
 {
     public function index()
     {
-        $tenues = Tenue::with('categorie')->get();
+        if (!auth()->user()->hasPermission('Gérer les produits')) {
+            abort(403, 'Accès interdit');
+        }
+        $tenues = Tenue::with('category')->get();
         return view('tenues.index', compact('tenues'));
     }
 
     public function create()
     {
-        $categories = Categorie::all();
+        if (!auth()->user()->hasPermission('Gérer les produits')) {
+            abort(403, 'Accès interdit');
+        }
+        $categories = Category::all();
         return view('tenues.create', compact('categories'));
     }
 
     public function store(TenueRequest $request)
     {
+        
+        if (!auth()->user()->hasPermission('Gérer les produits')) {
+            abort(403, 'Accès interdit');
+        }
         $data = $request->validated();
         if ($request->hasFile('image')) {
             $data['image'] = $request->file('image')->store('images', 'public');
@@ -33,17 +43,26 @@ class TenueController extends Controller
 
     public function show(Tenue $tenue)
     {
+        if (!auth()->user()->hasPermission('Gérer les produits')) {
+            abort(403, 'Accès interdit');
+        }
         return view('tenues.show', compact('tenue'));
     }
 
     public function edit(Tenue $tenue)
     {
-        $categories = Categorie::all();
+        if (!auth()->user()->hasPermission('Gérer les produits')) {
+            abort(403, 'Accès interdit');
+        }
+        $categories = Category::all();
         return view('tenues.edit', compact('tenue', 'categories'));
     }
 
     public function update(TenueRequest $request, Tenue $tenue)
     {
+        if (!auth()->user()->hasPermission('Gérer les produits')) {
+            abort(403, 'Accès interdit');
+        }
         $data = $request->validated();
         if ($request->hasFile('image')) {
             $data['image'] = $request->file('image')->store('images', 'public');
@@ -54,6 +73,9 @@ class TenueController extends Controller
 
     public function destroy(Tenue $tenue)
     {
+        if (!auth()->user()->hasPermission('Gérer les produits')) {
+            abort(403, 'Accès interdit');
+        }
         $tenue->delete();
         return redirect()->route('tenues.index')->with('success', 'Tenue supprimée avec succès.');
     }
